re-organize the tests util functions
